Split out utility files into a new subpackage 'gph_utils'
