+++ setup_contest.py
@@ -43,7 +43,7 @@ import argparse
 from hashlib import sha1
 from contests import *
 from gph_config import *
-from gph_logging import log_message
+from gph_utils.gph_logging import log_message
 import hs_wrapper as hs
 
 parser = argparse.ArgumentParser()
